Рефакторит компонент LocationItem
